+++ test/example/javascript/callback.js
@@ -1,0 +1,2 @@
+var a = call(function () {});
+const b = call(() => {});